Adds the initial support for AOTInductor model and interface
This PR combines the C++ code for the AOTInductor's model and interface with
Bin Bao's changes to AOTInductor codegen.

It adds a number of AOTInductor C interfaces that can be used by an
inference runtime. Under the hood of the interfaces, the model code generated by
the AOTInductor's codegen is wrapped into a class, AOTInductorModel, which
manages tensors and run the model inference.

On top of AOTInductorModel, we provide one more abstract layer,
AOTInductorModelContainer, which allows the user to have multiple inference
runs concurrently for the same model.

This PR also adjusts the compilation options for AOT codegen, particularly
some fbcode-related changes such as libs to be linked and header-file search
paths.

Note that this is the very first version of the AOTInductor model and interface,
so many features (e.g. dynamic shape) are incomplete. We will support those
missing features in in future PRs.

ghstack-source-id: 67c0e7de78a3f92cf09265586911d643430fa794
Pull Request resolved: https://github.com/pytorch/pytorch/pull/103922